requirements.txt
google-generativeai==0.3.2
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==2.6.1
jinja2==3.1.3
python-multipart==0.0.9 